cli: `validation create` command
