+++ cli/run.ts
@@ -1,9 +1,8 @@
 import { Command, CompletionsCommand } from 'cliffy/command/mod.ts';
 
 import authCommand from './auth_command.ts';
-import requestCommand from './request_command.ts';
 import submissionCommand from './submission_command.ts';
-import { submitCommand, validateCommand } from './database_commands.ts';
+import validationCommand from './validation_command.ts';
 
 type Options = {
   apiUrl?: string;
@@ -16,9 +15,7 @@ const main: Command<Options> = new Command<Options>()
   .env('DDBJ_REPOSITORY_API_URL=<url:string>', 'API endpoint URL', { global: true, prefix: 'DDBJ_REPOSITORY_' })
   .action(() => main.showHelp())
   .command('auth', authCommand)
-  .command('validate', validateCommand)
-  .command('submit', submitCommand)
-  .command('request', requestCommand)
+  .command('validation', validationCommand)
   .command('submission', submissionCommand)
   .command('completion', new CompletionsCommand())
   .reset();
